consistent indenting
code is easier to read when indented consistently. I've tried to stick to the style in place which looks like 2 spaces for indents, and `{` at the end of a line, and `else` also on its own new line.

you should see nice vertical lines where code starts and finishes: e.g. the beginning and end of the for loop, and each if/else block.